requirements.txt
Flask
pandas
spacy
python-docx